fix(data-warehouse): isolate DuckDB table rendering
