+++ posthog/hogql/database/test/test_s3_table_duckdb.py
@@ -8,10 +8,10 @@ import duckdb
 from parameterized import parameterized
 
 from posthog.hogql.context import HogQLContext
-from posthog.hogql.database.s3_table import (
+from posthog.hogql.database.s3_table import S3Table
+from posthog.hogql.database.s3_table_duckdb import (
     DuckDBAzureSource,
     DuckDBS3Source,
-    S3Table,
     parse_duckdb_azure_source,
     parse_duckdb_s3_source,
 )
